Astar logic
We cant remove from priority queue (for now) so we must check if
node is on OpenList or ClosedList with better score. Ignore node
on priority queue if so.
End only if node is not the final and reopen, so we need a little
"break" on the a_star main loop.