Coord use template, not std::vector. GIANT COMMIT
Some major changes on the Project, incluing optimization and new code standards
The logic of the program did not change, but almost every method must be
rewriten.

- Coord class change 'std::vector m_coord' to 'm_coord[N]' where N is template.
This means every class in this project must be rewriten to use templates and I
can't break this commit in smaller parts, because it won't compile.

The advantages on speed are clear:
3 Sequences:
Before:  0m9.513s
After:  0m5.153s

4 Sequences:
Before:  0m21.020s
After:  0m11.480s

5 Sequences:
Before:  3m11.770s
After:  1m53.397s

- Remove of the NO_LIB_BOOST
This code was used to make sure that the libboost code was correct, but now
the Priority list is so mature, don't makes sense to keep both implementations.
BTW, I dont want to port this slow code with high memory consume to template.

- 'using namespace std'
Also, decided to remove the 'using namespace std' from files.
On the beginning of the project, we have only a vector and map, so it was
pretty obvious. But now, we are mixing the libboost, where sometimes objects
have the same name. And now we are using the std::multithread space that I fell
it is better to always use the std:: prefix.
This is the new project standard: do not use 'using namespace std' and always
add the std:: prefix when necessary.