AffineGaps: cleanup code
The affine gaps logic was implemented but for some reason it never
worked for the MSA problem. I have many ideas, buy me a beer and I
will try to explain it to you. Although, all previous results are
valid because GapExtension and GapOpen had the same cost. Now the old
code was removed using and we are only using one gap cost logic. It
is a bit faster